+++ src/plone/restapi/setuphandlers.py
@@ -1,7 +1,10 @@
 # -*- coding: utf-8 -*-
+from Acquisition import aq_inner
+from Acquisition import aq_parent
 from plone.restapi.pas.plugin import JWTAuthenticationPlugin
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.interfaces import INonInstallable
+from Products.PluggableAuthService.interfaces.authservice import IPluggableAuthService  # noqa: E501
 from zope.interface import implementer
 
 
@@ -25,18 +28,20 @@ class HiddenProfiles(object):
 
 
 def install_pas_plugin(context):
-    uf = getToolByName(context, 'acl_users', None)
-    if uf is None:
-        return
-
-    if 'jwt_auth' not in uf:
-        plugin = JWTAuthenticationPlugin('jwt_auth')
-        uf._setObject(plugin.getId(), plugin)
-        plugin = uf['jwt_auth']
-        plugin.manage_activateInterfaces([
-            'IAuthenticationPlugin',
-            'IExtractionPlugin',
-        ])
+    uf_parent = aq_inner(context)
+    while True:
+        uf = getToolByName(uf_parent, 'acl_users')
+        if IPluggableAuthService.providedBy(uf) and 'jwt_auth' not in uf:
+            plugin = JWTAuthenticationPlugin('jwt_auth')
+            uf._setObject(plugin.getId(), plugin)
+            plugin = uf['jwt_auth']
+            plugin.manage_activateInterfaces([
+                'IAuthenticationPlugin',
+                'IExtractionPlugin',
+            ])
+        if uf_parent is uf_parent.getPhysicalRoot():
+            break
+        uf_parent = aq_parent(uf_parent)
 
 
 def import_various(context):