+++ src/plone/restapi/types/z3crelationadapter.py
@@ -1,12 +1,11 @@
 # -*- coding: utf-8 -*-
+from plone.restapi.types.adapters import ListJsonSchemaProvider
 from plone.restapi.types.interfaces import IJsonSchemaProvider
 from z3c.relationfield.interfaces import IRelationList
 from zope.component import adapter
 from zope.component import getMultiAdapter
 from zope.interface import implementer
 from zope.interface import Interface
-
-from plone.restapi.types.adapters import ListJsonSchemaProvider
 
 
 @adapter(IRelationList, Interface, Interface)
@@ -23,3 +22,6 @@ class ChoiceslessRelationListSchemaProvider(ListJsonSchemaProvider):
         value_type_adapter.should_render_choices = False
 
         return value_type_adapter.get_schema()
+
+    def get_widget(self):
+        return 'relation'