+++ src/plone/restapi/services/users/add.py
@@ -1,14 +1,16 @@
 # -*- coding: utf-8 -*-
 
 from AccessControl import getSecurityManager
-
 from plone.restapi.deserializer import json_body
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.services import Service
-from Products.CMFCore.utils import getToolByName
 from Products.CMFCore.permissions import AddPortalMember
 from Products.CMFCore.permissions import SetOwnPassword
+from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.RegistrationTool import get_member_by_login_name
+from Products.CMFPlone.utils import getFSVersionTuple
 from zope.component import getAdapter
+from zope.component import getMultiAdapter
 from zope.component import queryMultiAdapter
 from zope.component.hooks import getSite
 from zope.interface import alsoProvides
@@ -30,6 +32,8 @@ except ImportError:   # pragma: no cover
     # Plone 5.0 and earlier
     from Products.PasswordResetTool.PasswordResetTool import ExpiredRequestError  # noqa
     from Products.PasswordResetTool.PasswordResetTool import InvalidRequestError  # noqa
+
+PLONE5 = getFSVersionTuple()[0] >= 5
 
 
 @implementer(IPublishTraverse)
@@ -149,12 +153,43 @@ class UsersPost(Service):
         send_password_reset = data.pop('sendPasswordReset', None)
         properties = data
 
-        # set username based on the login settings (username or email)
-        if security.use_email_as_login:
-            username = email
-            properties['username'] = email
+        if PLONE5:
+            # We are improving the way the userid/login_name is generated using
+            # Plone's plone.app.users utilities directly. Plone 4 lacks of the
+            # login_name one, so we leave it as it is, improving the Plone 5
+            # story
+            user_id_login_name_data = {
+                'username': username,
+                'email': email,
+                'fullname': data.get('fullname', '')
+            }
+
+            register_view = getMultiAdapter(
+                (self.context, self.request),
+                name='register'
+            )
+
+            register_view.generate_user_id(user_id_login_name_data)
+            register_view.generate_login_name(user_id_login_name_data)
+
+            user_id = user_id_login_name_data.get(
+                'user_id',
+                data.get('username')
+            )
+            login_name = user_id_login_name_data.get(
+                'login_name',
+                data.get('username')
+            )
+
+            username = user_id
+            properties['username'] = user_id
         else:
-            properties['username'] = username
+            # set username based on the login settings (username or email)
+            if security.use_email_as_login:
+                username = email
+                properties['username'] = email
+            else:
+                properties['username'] = username
 
         properties['email'] = email
 
@@ -177,6 +212,16 @@ class UsersPost(Service):
                 type='MissingParameterError',
                 message=str(e)))
 
+        if PLONE5:
+            # After user creation, we have to fix the login_name if it differs.
+            # That happens when the email login is enabled and we are using
+            # UUID as user ID security settings.
+            if user_id != login_name:
+                # The user id differs from the login name.  Set the login
+                # name correctly.
+                pas = getToolByName(self.context, 'acl_users')
+                pas.updateLoginName(user_id, login_name)
+
         if send_password_reset:
             registration.registeredNotify(username)
         self.request.response.setStatus(201)
@@ -193,6 +238,10 @@ class UsersPost(Service):
         portal = getSite()
         portal_membership = getToolByName(portal, 'portal_membership')
         return portal_membership.getMemberById(user_id)
+
+    def _get_user_by_login_name(self, user_id):
+        return get_member_by_login_name(
+            self.context, user_id, raise_exceptions=False)
 
     def _error(self, status, type, message):
         self.request.response.setStatus(status)
@@ -217,7 +266,7 @@ class UsersPost(Service):
 
     def update_password(self, data):
         username = self.params[0]
-        target_user = self._get_user(username)
+        target_user = self._get_user_by_login_name(username)
         reset_token = data.get('reset_token', None)
         old_password = data.get('old_password', None)
         new_password = data.get('new_password', None)