+++ CHANGES.rst
@@ -83,6 +83,9 @@ New Features:
 
 Bugfixes:
 
+- Add the 'widget' property to the @types endpoint for IRelationList fields.
+  [thet]
+
 - Do not include HTTP examples using data_files anymore, but move them below
   src/plone/restapi instead and use package_data to include them.
   [lgraf]